x/truststore/types/message_create_attestation_test.go
package types

import (
	"testing"

	"github.com/blarsy/truststore/testutil/sample"
	sdkerrors "github.com/cosmos/cosmos-sdk/types/errors"
	"github.com/stretchr/testify/require"
)

func TestMsgCreateAttestation_ValidateBasic(t *testing.T) {
	tests := []struct {
		name string
		msg  MsgCreateAttestation
		err  error
	}{
		{
			name: "invalid address",
			msg: MsgCreateAttestation{
				Creator: "invalid_address",
			},
			err: sdkerrors.ErrInvalidAddress,
		}, {
			name: "valid address",
			msg: MsgCreateAttestation{
				Creator: sample.AccAddress(),
			},
		},
	}
	for _, tt := range tests {
		t.Run(tt.name, func(t *testing.T) {
			err := tt.msg.ValidateBasic()
			if tt.err != nil {
				require.ErrorIs(t, err, tt.err)
				return
			}
			require.NoError(t, err)
		})
	}
}
